chore: moving and deleting some specific files
