+++ src/main.ts
@@ -1,8 +1,13 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { ConfigService } from '@nestjs/config';
+import { AppModule } from 'src/modules';
+import { ServerConfigType, SERVER_CONFIG } from 'src/config'; 
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  await app.listen(15001);
+  const configService = app.get(ConfigService);
+  const { port } = configService.get<ServerConfigType>(SERVER_CONFIG);
+  await app.listen(port);
 }
+
 bootstrap();
